+++ packages/tsargp/package.json
@@ -1,6 +1,6 @@
 {
   "name": "tsargp",
-  "version": "1.11.1",
+  "version": "1.11.2",
   "author": {
     "name": "Diego Sogari",
     "email": "diego.sogari@gmail.com",
