+++ packages/tsargp/src/library/index.ts
@@ -1,7 +1,8 @@
-export * from './options.js';
-export * from './styles.js';
+export { allOf, not, oneOf, valuesFor } from './common.js';
 export * from './config.js';
 export * from './enums.js';
 export * from './formatter.js';
+export * from './options.js';
 export * from './parser.js';
+export * from './styles.js';
 export * from './validator.js';
